moved ComboBox the seperate file
